backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
python-dotenv
flask-smorest
marshmallow-sqlalchemy
passlib[bcrypt]
flask-jwt-extended